moved MainActivity to presentation layer
