--- app/src/main/java/com/example/WeatherWiseApp/presentation/Main.java
@@ -1,15 +1,0 @@
-package com.example.WeatherWiseApp.presentation;
-
-import com.example.WeatherWiseApp.logic.WeatherManager;
-import com.example.WeatherWiseApp.objects.City;
-
-public class Main {
-    public static void main(String[] args) {
-        //CityProvider cityProvider = new CityProvider();
-        //WeatherManager weatherManager = new WeatherManager();
-
-        //City defaultCity = new City(cityProvider.getCity());
-        //System.out.println("The default city is: " + defaultCity.getCity());
-        //System.out.println(weatherManager.getWeatherJSON(defaultCity));
-    }
-}
